feat: 저장된 게임이 진행 중이면 삭제하지 못하는 기능 추가

+++ src/main/java/chess/service/ChessService.java
@@ -113,7 +113,14 @@ public class ChessService {
     }
 
     public void deleteGameByGameID(String gameID) {
+        checkCanDelete(GameTurn.find(chessGameDao.findTurnByID(gameID)));
         chessGameDao.deleteByGameID(gameID);
         pieceDao.deleteByGameID(gameID);
+    }
+
+    private void checkCanDelete(GameTurn gameTurn) {
+        if (!GameTurn.FINISHED.equals(gameTurn)) {
+            throw new IllegalArgumentException("아직 게임이 진행 중이라구!! 😡");
+        }
     }
 }
